ref(crons): Rearrange timeline component structure
This change is a holistic view of how the cron timelines components
should be organize.

This changes makes the following distinctions

 - The generic timeline components are separate from the overview
   timeline (which makes use of timeline components). Thus, the generic
   components are broken into a `timeline` component module, while
   `overviewTimeline` remains.

 - Various utils which were only used for the timeline have been moved
   into components/timeline/utils.

 - The general catch-all `utils` file has been removed and it's contents
   were either inlined in their single-use call sites, or moved into
   their own modules.

 - Timeline related hooks now live in timeline/hooks

 - A few components have been renamed to better clarify their purpose

   - TimelineTableRow    -> OverviewRow
   - CronDetailsTimeline -> DetailsTimeline
   - TimelinePlaceholder -> CheckInPlaceholder
   - JobTickTooltip      -> CheckInTooltip
