+++ static/app/views/monitors/components/monitorCheckIns.tsx
@@ -11,7 +11,7 @@ import Pagination from 'sentry/components/pagination';
 import {PanelTable} from 'sentry/components/panels/panelTable';
 import Placeholder from 'sentry/components/placeholder';
 import ShortId from 'sentry/components/shortId';
-import StatusIndicator from 'sentry/components/statusIndicator';
+import StatusIndicator, {StatusIndicatorProps} from 'sentry/components/statusIndicator';
 import Text from 'sentry/components/text';
 import {Tooltip} from 'sentry/components/tooltip';
 import {IconDownload} from 'sentry/icons';
@@ -27,12 +27,22 @@ import {ContextType} from 'sentry/views/discover/table/quickContext/utils';
 import type {CheckIn, Monitor, MonitorEnvironment} from 'sentry/views/monitors/types';
 import {CheckInStatus} from 'sentry/views/monitors/types';
 import {statusToText} from 'sentry/views/monitors/utils';
-import {checkStatusToIndicatorStatus} from 'sentry/views/monitors/utils/constants';
 
 type Props = {
   monitor: Monitor;
   monitorEnvs: MonitorEnvironment[];
   orgSlug: string;
+};
+
+export const checkStatusToIndicatorStatus: Record<
+  CheckInStatus,
+  StatusIndicatorProps['status']
+> = {
+  [CheckInStatus.OK]: 'success',
+  [CheckInStatus.ERROR]: 'error',
+  [CheckInStatus.IN_PROGRESS]: 'muted',
+  [CheckInStatus.MISSED]: 'warning',
+  [CheckInStatus.TIMEOUT]: 'error',
 };
 
 function MonitorCheckIns({monitor, monitorEnvs, orgSlug}: Props) {
