+++ static/app/views/monitors/components/timeline/gridLines.tsx
@@ -7,11 +7,10 @@ import {updateDateTime} from 'sentry/actionCreators/pageFilters';
 import {DateTime} from 'sentry/components/dateTime';
 import {space} from 'sentry/styles/space';
 import useRouter from 'sentry/utils/useRouter';
-import type {TimeWindowConfig} from 'sentry/views/monitors/components/overviewTimeline/types';
 
 import {useTimelineCursor} from './timelineCursor';
 import {useTimelineZoom} from './timelineZoom';
-import {alignDateToBoundary} from './utils';
+import type {TimeWindowConfig} from './types';
 
 interface Props {
   timeWindowConfig: TimeWindowConfig;
@@ -44,6 +43,25 @@ interface TimeMarker {
    * The position in pixels of the tick
    */
   position: number;
+}
+
+/**
+ * Aligns the given date to the start of a unit (minute, hour, day) based on
+ * the minuteInterval size. This will align to the right side of the boundary
+ *
+ * 01:53:43 (10m interval) => 01:54:00
+ * 01:32:00 (2hr interval) => 02:00:00
+ */
+function alignDateToBoundary(date: moment.Moment, minuteInterval: number) {
+  if (minuteInterval < 60) {
+    return date.minute(date.minutes() - (date.minutes() % minuteInterval)).seconds(0);
+  }
+
+  if (minuteInterval < 60 * 24) {
+    return date.startOf('hour');
+  }
+
+  return date.startOf('day');
 }
 
 function getTimeMarkersFromConfig(config: TimeWindowConfig, width: number) {
