+++ static/app/views/monitors/components/timeline/timelineCursor.tsx
