+++ static/app/views/monitors/components/timeline/utils/getAggregateStatusFromMultipleBuckets.tsx
@@ -1,4 +1,4 @@
-import type {MonitorBucketEnvMapping} from 'sentry/views/monitors/components/overviewTimeline/types';
+import type {MonitorBucketEnvMapping} from '../types';
 
 import {CHECKIN_STATUS_PRECEDENT} from './constants';
 import {getAggregateStatus} from './getAggregateStatus';
