+++ static/app/views/monitors/components/timeline/utils/getTimeRangeFromEvent.spec.tsx
