+++ static/app/views/monitors/components/timeline/utils/isEnvMappingEmpty.tsx
@@ -1,4 +1,4 @@
-import type {MonitorBucketEnvMapping} from 'sentry/views/monitors/components/overviewTimeline/types';
+import type {MonitorBucketEnvMapping} from '../types';
 
 /**
  * Determines if an environment mapping includes any job run data
